+++ docs/guide/security.md
@@ -2,6 +2,8 @@
 
 bQuery sanitizes DOM writes by default and supports Trusted Types. Use the security module directly when you need explicit control over sanitization and CSP helpers.
 
+As of 1.2.x, the security module is internally modularized (sanitize core, Trusted Types, CSP helpers, and constants). The **public API remains unchanged**, so you continue to import from `@bquery/bquery/security` as usual.
+
 ```ts
 import { sanitize, escapeHtml, stripTags } from '@bquery/bquery/security';
 
@@ -63,6 +65,7 @@ bQuery automatically adds `rel="noopener noreferrer"` to links that:
 2. Point to external domains (different origin)
 
 This protects against:
+
 - **Tabnabbing attacks**: prevents the opened page from accessing `window.opener`
 - **Referrer leakage**: prevents sensitive information in the URL from being sent
